+++ doc/version_history.rst
@@ -6,6 +6,15 @@
 Version History
 ##################
 
+.. _lsst.ts.mtdomegui-0.6.2:
+
+-------------
+0.6.2
+-------------
+
+* Enabled the statusLLC command.
+* Use the enums of **ControlMode** and **Brake** in **ts_xml**.
+
 .. _lsst.ts.mtdomegui-0.6.1:
 
 -------------